Refactored all files in js, created new subdirectory called js_function wherein we will put all of the js functions
