+++ app/checkout.html
@@ -149,7 +149,7 @@
         // var property = document.getElementById("completedBtn1");
         // property.style.backgroundColor = '#7FFF00';
 
-        require('./js/checkout_functions.js')
+        require('./js/js_functions/checkout_functions.js')
         require('./js/checkout.js')
 
     </script>
