+++ app/createteam.html
@@ -20,6 +20,7 @@
         <li><input type="button" name="cancel" value="Cancel" id="cancelBtn"></li>
     </ul>
 
+    <script src="js/js_functions/createteam_functions.js"></script>
     <script src="./js/createteam.js"></script>
 </body>
