+++ app/dashboard.html
@@ -18,6 +18,7 @@
     <p>Number of visual studio processes open</p>
     <p id="vscode-num"></p>
     <script>
+        require('./js/js_functions/dashboard_functions.js')
         require('./js/dashboard.js')
     </script>
 </body>
