+++ app/googlesignin.html
@@ -14,6 +14,7 @@
 
 <body>
     <h3 id="loginstatus">Logging in..</h3>
+    <script src="./js/js_functions/googlesignin_functions.js"></script>
     <script src="./js/googlesignin.js"></script>
 </body>
