+++ app/js/createteam.js
@@ -42,4 +42,3 @@ joinTeamButton.addEventListener('click', () => {
 
 var cancelButton = document.getElementById("cancelBtn")
 cancelButton.addEventListener("click", () => cancel())
-function cancel() { document.location.href = "taskbar.html" }
