+++ app/js/jointeam.js
@@ -55,4 +55,3 @@ joinTeamButton.addEventListener("click", function() {
 
 var cancelButton = document.getElementById("cancelBtn")
 cancelButton.addEventListener("click", () => cancel())
-function cancel() { document.location.href = "taskbar.html" }
