+++ app/js/js_functions/checkin_functions.js
