+++ app/js/js_functions/checkout_functions.js
