+++ app/js/js_functions/createteam_functions.js
@@ -1,0 +1,3 @@
+function cancel() { document.location.href = "taskbar.html" }
+
+module.exports = {cancel};
