+++ app/js/js_functions/googlesignin_functions.js
@@ -1,0 +1,1 @@
+//Currently no functions
