+++ app/js/js_functions/signup_functions.js
@@ -1,0 +1,1 @@
+//No functions currently
