+++ app/signup.html
@@ -24,6 +24,7 @@
     <li>Already have an account? <a href="signin.html">Sign In</a></li>
   </ul>
 
+  <script> require('./js/js_functions/signup_functions.js') </script>
   <script> require('./js/signup.js') </script>
 </body>
