+++ app/taskbar.html
@@ -63,6 +63,7 @@
             </ul>
         </div>
     </div>
+    <script src="./js/js_functions/taskbar_functions.js"></script>
     <script src="./js/taskbar.js"></script>
 </body>
