+++ jscoverage/lcov-report/sum.js.html
@@ -3,7 +3,7 @@
 <html lang="en">
 
 <head>
-    <title>Code coverage report for sum.js</title>
+    <title>Code coverage report for sum_functions.js</title>
     <meta charset="utf-8" />
     <link rel="stylesheet" href="prettify.css" />
     <link rel="stylesheet" href="base.css" />
@@ -19,7 +19,7 @@
 <body>
 <div class='wrapper'>
     <div class='pad1'>
-        <h1><a href="index.html">All files</a> sum.js</h1>
+        <h1><a href="index.html">All files</a> sum_functions.js</h1>
         <div class='clearfix'>
             
             <div class='fl pad1y space-right2'>
@@ -95,7 +95,7 @@
 <span class="cline-any cline-neutral">&nbsp;</span>
 <span class="cline-any cline-neutral">&nbsp;</span>
 <span class="cline-any cline-neutral">&nbsp;</span>
-<span class="cline-any cline-neutral">&nbsp;</span></td><td class="text"><pre class="prettyprint lang-js">/* sum.js */
+<span class="cline-any cline-neutral">&nbsp;</span></td><td class="text"><pre class="prettyprint lang-js">/* sum_functions.js */
 &nbsp;
 module.exports = function() {
 &nbsp;
