+++ jscoverage/lcov-report/test/unit_testing/index.html
@@ -73,7 +73,7 @@
 </tr>
 </thead>
 <tbody><tr>
-	<td class="file high" data-value="sum.js"><a href="sum.js.html">sum.js</a></td>
+	<td class="file high" data-value="sum_functions.js"><a href="sum.js.html">sum_functions.js</a></td>
 	<td data-value="100" class="pic high">
 	<div class="chart"><div class="cover-fill cover-full" style="width: 100%"></div><div class="cover-empty" style="width: 0%"></div></div>
 	</td>
